% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map the locations of all accidents in a state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{integer State number as seen in FARS dataset}

\item{year}{a scalar integer value}
}
\description{
Plot/Map the locations where accidents occurred
in the specified year for the specified state.
No plot is made if data does not exist for the specified state and year.
Error if the state has no data for the specified year.
}
\examples{
\dontrun{
fars_map_state(39, 2014)
}
}
