% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS data from (compressed) csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a string value specifying path to csv file}
}
\value{
tibble with observations from FARS csv file.
Error if file not found.
}
\description{
Read FARS data from (compressed) csv file
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
}
}
