% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Retrieve all months (by year, year must be present in
 \code{years} argument) for which an observation exists}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of integer values}
}
\value{
list of tibbles with only month and year variables
for all observations found.
List item is NULL if csv file not found for a particular year.
}
\description{
Retrieve all months (by year, year must be present in
 \code{years} argument) for which an observation exists
}
\examples{
\dontrun{
fars_read_years(2013:2017)
}
}
