% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize observations by month across years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of integer values}
}
\value{
tibble count of observations for all the months across years.
}
\description{
Summarize observations by month across years
}
\examples{
fars_summarize_years(2013:2014)

}
