% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Construct csv filename using a specific \code{year} argument}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a scalar integer value}
}
\value{
FARS csv filename corresponding to the \code{year}
}
\description{
Construct csv filename using a specific \code{year} argument
}
\examples{
make_filename(2013)
make_filename(2015)

}
